MovableTypeRebuilder/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
